refactor(app_partner): split settlement_page.dart — extract dashboard/list tabs into part files
Splits 11 private sub-classes out of the 535-line settlement_page.dart
into two `part of` files:
- _settlement_dashboard_tab.dart: _DashboardTab, _PeriodSelector,
  _RevenueSummaryCard, _StatusSummaryGrid, _StatusCell
- _settlement_list_tab.dart: _ListItem, _MonthHeader, _CardItem,
  _ListTab, _ListTabState, _MonthHeaderWidget

Main file retains only SettlementPage + _SettlementPageState (~70 lines).

Constraint: part-of pattern required — all sub-classes are private (_), cannot be moved to separate library files
Confidence: high
Scope-risk: narrow

Closes #2146
